Fix signed/unsigned warning in the split function.
The Windows compiler correctly pointed out a signed/unsigned
mismatch in the split function, in that we were storing a negative
number in a size_t (I'm not sure why gcc/clang didn't show this,
but oh well).  Slightly rewrite the split function to not use
negative numbers, which makes things happier.  Note that I also
changed the name of the function here to "split", since I'll be
re-using this function again for other purposes.