A bunch of fixes to please static analysis.
clang static analysis points out a number of problems in the
tests.  A few of these are real; in the unlikely event that
they ever failed, the test would crash.  The rest are false
positives in that the test as currently written would prevent
the crash, but clang throws a warning anyway.  However, we
can please clang and keep the original intent by switching
most of these to ASSERT_NE(nullptr, ptr_to_test), which will
have better error messages anyway.  With this change, sdformat
is clang static analysis clean.